@@ -1,78 +1,97 @@
+# frozen_string_literal: true
+
 module Faraday
   class Adapter
+    # HTTPClient adapter.
     class HTTPClient < Faraday::Adapter
       dependency 'httpclient'
 
-      def client
-        @client ||= ::HTTPClient.new
+      def build_connection(env)
+        @client ||= ::HTTPClient.new.tap do |cli|
+          # enable compression
+          cli.transparent_gzip_decompression = true
+        end
+
+        if (req = env[:request])
+          if (proxy = req[:proxy])
+            configure_proxy @client, proxy
+          end
+
+          if (bind = req[:bind])
+            configure_socket @client, bind
+          end
+
+          configure_timeouts @client, req
+        end
+
+        if env[:url].scheme == 'https' && (ssl = env[:ssl])
+          configure_ssl @client, ssl
+        end
+
+        configure_client @client
+
+        @client
       end
 
       def call(env)
         super
 
-        # enable compression
-        client.transparent_gzip_decompression = true
-
-        if req = env[:request]
-          if proxy = req[:proxy]
-            configure_proxy proxy
-          end
-
-          if bind = req[:bind]
-            configure_socket bind
-          end
-
-          configure_timeouts req
-        end
-
-        if env[:url].scheme == 'https' && ssl = env[:ssl]
-          configure_ssl ssl
-        end
-
-        configure_client
-
-        # TODO Don't stream yet.
+        # TODO: Don't stream yet.
         # https://github.com/nahi/httpclient/pull/90
         env[:body] = env[:body].read if env[:body].respond_to? :read
 
-        resp = client.request env[:method], env[:url],
-          :body   => env[:body],
-          :header => env[:request_headers]
+        connection(env) do |http|
+          resp = http.request env[:method], env[:url],
+                              body: env[:body],
+                              header: env[:request_headers]
 
-        save_response env, resp.status, resp.body, resp.headers, resp.reason
+          if (req = env[:request]).stream_response?
+            warn "Streaming downloads for #{self.class.name} " \
+              'are not yet implemented.'
+            req.on_data.call(resp.body, resp.body.bytesize)
+          end
+          save_response env, resp.status, resp.body, resp.headers, resp.reason
 
-        @app.call env
+          @app.call env
+        end
       rescue ::HTTPClient::TimeoutError, Errno::ETIMEDOUT
-        raise Faraday::Error::TimeoutError, $!
-      rescue ::HTTPClient::BadResponseError => err
-        if err.message.include?('status 407')
-          raise Faraday::Error::ConnectionFailed, %{407 "Proxy Authentication Required "}
-        else
-          raise Faraday::Error::ClientError, $!
+        raise Faraday::TimeoutError, $ERROR_INFO
+      rescue ::HTTPClient::BadResponseError => e
+        if e.message.include?('status 407')
+          raise Faraday::ConnectionFailed,
+                %(407 "Proxy Authentication Required ")
         end
-      rescue Errno::ECONNREFUSED, IOError, SocketError
-        raise Faraday::Error::ConnectionFailed, $!
-      rescue => err
-        if defined?(OpenSSL) && OpenSSL::SSL::SSLError === err
-          raise Faraday::SSLError, err
-        else
-          raise
+
+        raise Faraday::ClientError, $ERROR_INFO
+      rescue Errno::EADDRNOTAVAIL, Errno::ECONNREFUSED, IOError, SocketError
+        raise Faraday::ConnectionFailed, $ERROR_INFO
+      rescue StandardError => e
+        if defined?(::OpenSSL::SSL::SSLError) && \
+           e.is_a?(::OpenSSL::SSL::SSLError)
+          raise Faraday::SSLError, e
         end
+
+        raise
       end
 
-      def configure_socket(bind)
+      # @param bind [Hash]
+      def configure_socket(client, bind)
         client.socket_local.host = bind[:host]
         client.socket_local.port = bind[:port]
       end
 
-      def configure_proxy(proxy)
+      # Configure proxy URI and any user credentials.
+      #
+      # @param proxy [Hash]
+      def configure_proxy(client, proxy)
         client.proxy = proxy[:uri]
-        if proxy[:user] && proxy[:password]
-          client.set_proxy_auth proxy[:user], proxy[:password]
-        end
+        return unless proxy[:user] && proxy[:password]
+
+        client.set_proxy_auth(proxy[:user], proxy[:password])
       end
 
-      def configure_ssl(ssl)
+      # @param ssl [Hash]
+      def configure_ssl(client, ssl)
         ssl_config = client.ssl_config
         ssl_config.verify_mode = ssl_verify_mode(ssl)
         ssl_config.cert_store = ssl_cert_store(ssl)
@@ -84,40 +103,45 @@ module Faraday
         ssl_config.verify_depth = ssl[:verify_depth] if ssl[:verify_depth]
       end
 
-      def configure_timeouts(req)
-        if req[:timeout]
-          client.connect_timeout   = req[:timeout]
-          client.receive_timeout   = req[:timeout]
-          client.send_timeout      = req[:timeout]
+      # @param req [Hash]
+      def configure_timeouts(client, req)
+        if (sec = request_timeout(:open, req))
+          client.connect_timeout = sec
         end
 
-        if req[:open_timeout]
-          client.connect_timeout   = req[:open_timeout]
-          client.send_timeout      = req[:open_timeout]
+        if (sec = request_timeout(:write, req))
+          client.send_timeout = sec
         end
+
+        return unless (sec = request_timeout(:read, req))
+
+        client.receive_timeout = sec
       end
 
-      def configure_client
-        @config_block.call(client) if @config_block
+      def configure_client(client)
+        @config_block&.call(client)
       end
 
+      # @param ssl [Hash]
+      # @return [OpenSSL::X509::Store]
       def ssl_cert_store(ssl)
         return ssl[:cert_store] if ssl[:cert_store]
+
         # Memoize the cert store so that the same one is passed to
-        # HTTPClient each time, to avoid resyncing SSL sesions when
+        # HTTPClient each time, to avoid resyncing SSL sessions when
         # it's changed
-        @cert_store ||= begin
+        @ssl_cert_store ||= begin
           # Use the default cert store by default, i.e. system ca certs
-          cert_store = OpenSSL::X509::Store.new
-          cert_store.set_default_paths
-          cert_store
+          OpenSSL::X509::Store.new.tap(&:set_default_paths)
         end
       end
 
+      # @param ssl [Hash]
       def ssl_verify_mode(ssl)
         ssl[:verify_mode] || begin
           if ssl.fetch(:verify, true)
-            OpenSSL::SSL::VERIFY_PEER | OpenSSL::SSL::VERIFY_FAIL_IF_NO_PEER_CERT
+            OpenSSL::SSL::VERIFY_PEER |
+              OpenSSL::SSL::VERIFY_FAIL_IF_NO_PEER_CERT
           else
             OpenSSL::SSL::VERIFY_NONE
           end
