@@ -1,81 +1,123 @@
+# frozen_string_literal: true
+
 module Faraday
   class Adapter
+    # Excon adapter.
     class Excon < Faraday::Adapter
       dependency 'excon'
+
+      def build_connection(env)
+        opts = opts_from_env(env)
+        ::Excon.new(env[:url].to_s, opts.merge(@connection_options))
+      end
 
       def call(env)
         super
 
-        opts = {}
-        if env[:url].scheme == 'https' && ssl = env[:ssl]
-          opts[:ssl_verify_peer] = !!ssl.fetch(:verify, true)
-          opts[:ssl_ca_path] = ssl[:ca_path] if ssl[:ca_path]
-          opts[:ssl_ca_file] = ssl[:ca_file] if ssl[:ca_file]
-          opts[:client_cert] = ssl[:client_cert] if ssl[:client_cert]
-          opts[:client_key]  = ssl[:client_key]  if ssl[:client_key]
-          opts[:certificate] = ssl[:certificate] if ssl[:certificate]
-          opts[:private_key] = ssl[:private_key] if ssl[:private_key]
-          opts[:ssl_version] = ssl[:version] if ssl[:version]
-          opts[:ssl_min_version] = ssl[:min_version] if ssl[:min_version]
-          opts[:ssl_max_version] = ssl[:max_version] if ssl[:max_version]
+        req_opts = {
+          method: env[:method].to_s.upcase,
+          headers: env[:request_headers],
+          body: read_body(env)
+        }
 
-          # https://github.com/geemus/excon/issues/106
-          # https://github.com/jruby/jruby-ossl/issues/19
-          opts[:nonblock] = false
-        end
-
-        if ( req = env[:request] )
-          if req[:timeout]
-            opts[:read_timeout]      = req[:timeout]
-            opts[:connect_timeout]   = req[:timeout]
-            opts[:write_timeout]     = req[:timeout]
-          end
-
-          if req[:open_timeout]
-            opts[:connect_timeout]   = req[:open_timeout]
-          end
-
-          if req[:proxy]
-            opts[:proxy] = {
-              :host     => req[:proxy][:uri].host,
-              :hostname => req[:proxy][:uri].hostname,
-              :port     => req[:proxy][:uri].port,
-              :scheme   => req[:proxy][:uri].scheme,
-              :user     => req[:proxy][:user],
-              :password => req[:proxy][:password]
-            }
+        req = env[:request]
+        if req&.stream_response?
+          total = 0
+          req_opts[:response_block] = lambda do |chunk, _remain, _total|
+            req.on_data.call(chunk, total += chunk.size)
           end
         end
 
-        conn = create_connection(env, opts)
+        resp = connection(env) { |http| http.request(req_opts) }
+        save_response(env, resp.status.to_i, resp.body, resp.headers,
+                      resp.reason_phrase)
 
-        resp = conn.request \
-          :method  => env[:method].to_s.upcase,
-          :headers => env[:request_headers],
-          :body    => read_body(env)
+        @app.call(env)
+      rescue ::Excon::Errors::SocketError => e
+        raise Faraday::TimeoutError, e if e.message.match?(/\btimeout\b/)
 
-        save_response(env, resp.status.to_i, resp.body, resp.headers, resp.reason_phrase)
+        raise Faraday::SSLError, e if e.message.match?(/\bcertificate\b/)
 
-        @app.call env
-      rescue ::Excon::Errors::SocketError => err
-        if err.message =~ /\btimeout\b/
-          raise Error::TimeoutError, err
-        elsif err.message =~ /\bcertificate\b/
-          raise Faraday::SSLError, err
-        else
-          raise Error::ConnectionFailed, err
-        end
-      rescue ::Excon::Errors::Timeout => err
-        raise Error::TimeoutError, err
-      end
-
-      def create_connection(env, opts)
-        ::Excon.new(env[:url].to_s, opts.merge(@connection_options))
+        raise Faraday::ConnectionFailed, e
+      rescue ::Excon::Errors::Timeout => e
+        raise Faraday::TimeoutError, e
       end
 
       # TODO: support streaming requests
       def read_body(env)
         env[:body].respond_to?(:read) ? env[:body].read : env[:body]
+      end
+
+      private
+
+      def opts_from_env(env)
+        opts = {}
+        amend_opts_with_ssl!(opts, env[:ssl]) if needs_ssl_settings?(env)
+
+        if (req = env[:request])
+          amend_opts_with_timeouts!(opts, req)
+          amend_opts_with_proxy_settings!(opts, req)
+        end
+
+        opts
+      end
+
+      def needs_ssl_settings?(env)
+        env[:url].scheme == 'https' && env[:ssl]
+      end
+
+      OPTS_KEYS = [
+        %i[client_cert client_cert],
+        %i[client_key client_key],
+        %i[certificate certificate],
+        %i[private_key private_key],
+        %i[ssl_ca_path ca_path],
+        %i[ssl_ca_file ca_file],
+        %i[ssl_version version],
+        %i[ssl_min_version min_version],
+        %i[ssl_max_version max_version]
+      ].freeze
+
+      def amend_opts_with_ssl!(opts, ssl)
+        opts[:ssl_verify_peer] = !!ssl.fetch(:verify, true)
+        # https://github.com/geemus/excon/issues/106
+        # https://github.com/jruby/jruby-ossl/issues/19
+        opts[:nonblock] = false
+
+        OPTS_KEYS.each do |(key_in_opts, key_in_ssl)|
+          next unless ssl[key_in_ssl]
+
+          opts[key_in_opts] = ssl[key_in_ssl]
+        end
+      end
+
+      def amend_opts_with_timeouts!(opts, req)
+        if (sec = request_timeout(:read, req))
+          opts[:read_timeout] = sec
+        end
+
+        if (sec = request_timeout(:write, req))
+          opts[:write_timeout] = sec
+        end
+
+        return unless (sec = request_timeout(:open, req))
+
+        opts[:connect_timeout] = sec
+      end
+
+      def amend_opts_with_proxy_settings!(opts, req)
+        opts[:proxy] = proxy_settings_for_opts(req[:proxy]) if req[:proxy]
+      end
+
+      def proxy_settings_for_opts(proxy)
+        {
+          host: proxy[:uri].host,
+          hostname: proxy[:uri].hostname,
+          port: proxy[:uri].port,
+          scheme: proxy[:uri].scheme,
+          user: proxy[:user],
+          password: proxy[:password]
+        }
       end
     end
   end
