@@ -1,16 +1,21 @@
+# frozen_string_literal: true
+
 begin
   require 'net/https'
 rescue LoadError
-  warn "Warning: no such file to load -- net/https. Make sure openssl is installed if you want ssl support"
+  warn 'Warning: no such file to load -- net/https. ' \
+    'Make sure openssl is installed if you want ssl support'
   require 'net/http'
 end
 require 'zlib'
 
 module Faraday
   class Adapter
+    # Net::HTTP adapter.
     class NetHttp < Faraday::Adapter
-      NET_HTTP_EXCEPTIONS = [
+      exceptions = [
         IOError,
+        Errno::EADDRNOTAVAIL,
         Errno::ECONNABORTED,
         Errno::ECONNREFUSED,
         Errno::ECONNRESET,
@@ -22,54 +27,78 @@ module Faraday
         Net::HTTPHeaderSyntaxError,
         Net::ProtocolError,
         SocketError,
-        Zlib::GzipFile::Error,
+        Zlib::GzipFile::Error
       ]
 
-      NET_HTTP_EXCEPTIONS << OpenSSL::SSL::SSLError if defined?(OpenSSL)
-      NET_HTTP_EXCEPTIONS << Net::OpenTimeout if defined?(Net::OpenTimeout)
+      if defined?(::OpenSSL::SSL::SSLError)
+        exceptions << ::OpenSSL::SSL::SSLError
+      end
+      exceptions << ::Net::OpenTimeout if defined?(::Net::OpenTimeout)
+
+      NET_HTTP_EXCEPTIONS = exceptions.freeze
 
       def initialize(app = nil, opts = {}, &block)
-        @cert_store = nil
+        @ssl_cert_store = nil
         super(app, opts, &block)
+      end
+
+      def build_connection(env)
+        net_http_connection(env).tap do |http|
+          if http.respond_to?(:use_ssl=)
+            http.use_ssl = env[:url].scheme == 'https'
+          end
+          configure_ssl(http, env[:ssl])
+          configure_request(http, env[:request])
+        end
+      end
+
+      def net_http_connection(env)
+        klass = if (proxy = env[:request][:proxy])
+                  Net::HTTP::Proxy(proxy[:uri].hostname, proxy[:uri].port,
+                                   proxy[:user], proxy[:password])
+                else
+                  Net::HTTP
+                end
+        port = env[:url].port || (env[:url].scheme == 'https' ? 443 : 80)
+        klass.new(env[:url].hostname, port)
       end
 
       def call(env)
         super
-        with_net_http_connection(env) do |http|
-          configure_ssl(http, env[:ssl]) if env[:url].scheme == 'https' and env[:ssl]
-          configure_request(http, env[:request])
-
+        http_response = connection(env) do |http|
           begin
-            http_response = perform_request(http, env)
-          rescue *NET_HTTP_EXCEPTIONS => err
-            if defined?(OpenSSL) && OpenSSL::SSL::SSLError === err
-              raise Faraday::SSLError, err
-            else
-              raise Error::ConnectionFailed, err
+            perform_request(http, env)
+          rescue *NET_HTTP_EXCEPTIONS => e
+            if defined?(OpenSSL) && e.is_a?(OpenSSL::SSL::SSLError)
+              raise Faraday::SSLError, e
             end
-          end
 
-          save_response(env, http_response.code.to_i, http_response.body || '', nil, http_response.message) do |response_headers|
-            http_response.each_header do |key, value|
-              response_headers[key] = value
-            end
+            raise Faraday::ConnectionFailed, e
+          end
+        end
+
+        save_response(env, http_response.code.to_i,
+                      http_response.body || +'', nil,
+                      http_response.message) do |response_headers|
+          http_response.each_header do |key, value|
+            response_headers[key] = value
           end
         end
 
         @app.call env
-      rescue Timeout::Error, Errno::ETIMEDOUT => err
-        raise Faraday::Error::TimeoutError, err
+      rescue Timeout::Error, Errno::ETIMEDOUT => e
+        raise Faraday::TimeoutError, e
       end
 
       private
 
       def create_request(env)
         request = Net::HTTPGenericRequest.new \
-          env[:method].to_s.upcase,    # request method
-          !!env[:body],                # is there request body
-          :head != env[:method],       # is there response body
-          env[:url].request_uri,       # request uri path
-          env[:request_headers]        # request headers
+          env[:method].to_s.upcase, # request method
+          !!env[:body], # is there request body
+          env[:method] != :head, # is there response body
+          env[:url].request_uri, # request uri path
+          env[:request_headers] # request headers
 
         if env[:body].respond_to?(:read)
           request.body_stream = env[:body]
@@ -80,62 +109,100 @@ module Faraday
       end
 
       def perform_request(http, env)
-        if :get == env[:method] and !env[:body]
-          # prefer `get` to `request` because the former handles gzip (ruby 1.9)
-          http.get env[:url].request_uri, env[:request_headers]
+        if env[:request].stream_response?
+          size = 0
+          yielded = false
+          http_response = request_with_wrapped_block(http, env) do |chunk|
+            if chunk.bytesize.positive? || size.positive?
+              yielded = true
+              size += chunk.bytesize
+              env[:request].on_data.call(chunk, size)
+            end
+          end
+          env[:request].on_data.call(+'', 0) unless yielded
+          # Net::HTTP returns something,
+          # but it's not meaningful according to the docs.
+          http_response.body = nil
+          http_response
         else
-          http.request create_request(env)
+          request_with_wrapped_block(http, env)
         end
       end
 
-      def with_net_http_connection(env)
-        yield net_http_connection(env)
+      def request_with_wrapped_block(http, env, &block)
+        if (env[:method] == :get) && !env[:body]
+          # prefer `get` to `request` because the former handles gzip (ruby 1.9)
+          request_via_get_method(http, env, &block)
+        else
+          request_via_request_method(http, env, &block)
+        end
       end
 
-      def net_http_connection(env)
-        if proxy = env[:request][:proxy]
-          Net::HTTP::Proxy(proxy[:uri].hostname, proxy[:uri].port, proxy[:user], proxy[:password])
-        else
-          Net::HTTP
-        end.new(env[:url].hostname, env[:url].port || (env[:url].scheme == 'https' ? 443 : 80))
+      def request_via_get_method(http, env, &block)
+        # Must use Net::HTTP#start and pass it a block otherwise the server's
+        # TCP socket does not close correctly.
+        http.start do |opened_http|
+          opened_http.get env[:url].request_uri, env[:request_headers], &block
+        end
+      end
+
+      def request_via_request_method(http, env, &block)
+        # Must use Net::HTTP#start and pass it a block otherwise the server's
+        # TCP socket does not close correctly.
+        http.start do |opened_http|
+          if block_given?
+            opened_http.request create_request(env) do |response|
+              response.read_body(&block)
+            end
+          else
+            opened_http.request create_request(env)
+          end
+        end
       end
 
       def configure_ssl(http, ssl)
-        http.use_ssl      = true
-        http.verify_mode  = ssl_verify_mode(ssl)
-        http.cert_store   = ssl_cert_store(ssl)
+        return unless ssl
 
-        http.cert         = ssl[:client_cert]  if ssl[:client_cert]
-        http.key          = ssl[:client_key]   if ssl[:client_key]
-        http.ca_file      = ssl[:ca_file]      if ssl[:ca_file]
-        http.ca_path      = ssl[:ca_path]      if ssl[:ca_path]
+        http.verify_mode = ssl_verify_mode(ssl)
+        http.cert_store = ssl_cert_store(ssl)
+
+        http.cert = ssl[:client_cert] if ssl[:client_cert]
+        http.key = ssl[:client_key] if ssl[:client_key]
+        http.ca_file = ssl[:ca_file] if ssl[:ca_file]
+        http.ca_path = ssl[:ca_path] if ssl[:ca_path]
         http.verify_depth = ssl[:verify_depth] if ssl[:verify_depth]
-        http.ssl_version  = ssl[:version]      if ssl[:version]
-        http.min_version  = ssl[:min_version]  if ssl[:min_version]
-        http.max_version  = ssl[:max_version]  if ssl[:max_version]
+        http.ssl_version = ssl[:version] if ssl[:version]
+        http.min_version = ssl[:min_version] if ssl[:min_version]
+        http.max_version = ssl[:max_version] if ssl[:max_version]
       end
 
       def configure_request(http, req)
-        if req[:timeout]
-          http.read_timeout  = req[:timeout]
-          http.open_timeout  = req[:timeout]
-          http.write_timeout = req[:timeout] if http.respond_to?(:write_timeout=)
+        if (sec = request_timeout(:read, req))
+          http.read_timeout = sec
         end
-        http.open_timeout  = req[:open_timeout]  if req[:open_timeout]
-        http.write_timeout = req[:write_timeout] if req[:write_timeout] && http.respond_to?(:write_timeout=)
-          # Only set if Net::Http supports it, since Ruby 2.5.
-        http.max_retries  = 0                    if http.respond_to?(:max_retries=)
 
-        @config_block.call(http) if @config_block
+        if (sec = http.respond_to?(:write_timeout=) &&
+                  request_timeout(:write, req))
+          http.write_timeout = sec
+        end
+
+        if (sec = request_timeout(:open, req))
+          http.open_timeout = sec
+        end
+
+        # Only set if Net::Http supports it, since Ruby 2.5.
+        http.max_retries = 0 if http.respond_to?(:max_retries=)
+
+        @config_block&.call(http)
       end
 
       def ssl_cert_store(ssl)
         return ssl[:cert_store] if ssl[:cert_store]
-        return @cert_store if @cert_store
-        # Use the default cert store by default, i.e. system ca certs
-        @cert_store = OpenSSL::X509::Store.new
-        @cert_store.set_default_paths
-        @cert_store
+
+        @ssl_cert_store ||= begin
+          # Use the default cert store by default, i.e. system ca certs
+          OpenSSL::X509::Store.new.tap(&:set_default_paths)
+        end
       end
 
       def ssl_verify_mode(ssl)
