@@ -1,37 +1,21 @@
+# frozen_string_literal: true
+
 module Faraday
+  # Middleware is the basic base class of any Faraday middleware.
   class Middleware
     extend MiddlewareRegistry
-
-    class << self
-      attr_accessor :load_error
-      private :load_error=
-    end
-
-    self.load_error = nil
-
-    # Executes a block which should try to require and reference dependent libraries
-    def self.dependency(lib = nil)
-      lib ? require(lib) : yield
-    rescue LoadError, NameError => error
-      self.load_error = error
-    end
-
-    def self.new(*)
-      raise "missing dependency for #{self}: #{load_error.message}" unless loaded?
-      super
-    end
-
-    def self.loaded?
-      load_error.nil?
-    end
-
-    def self.inherited(subclass)
-      super
-      subclass.send(:load_error=, self.load_error)
-    end
+    extend DependencyLoader
 
     def initialize(app = nil)
       @app = app
+    end
+
+    def close
+      if @app.respond_to?(:close)
+        @app.close
+      else
+        warn "#{@app} does not implement \#close!"
+      end
     end
   end
 end
